+++ WDL/blowfish.h
@@ -10,13 +10,13 @@ extern "C" {
 #endif
 
 typedef struct {
-  unsigned long P[16 + 2];
-  unsigned long S[4*256];
+  unsigned int P[16 + 2];
+  unsigned int S[4*256];
 } BLOWFISH_CTX;
 
 void Blowfish_Init(BLOWFISH_CTX *ctx, unsigned char *key, int keyLen);
-void Blowfish_Encrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr);
-void Blowfish_Decrypt(BLOWFISH_CTX *ctx, unsigned long *xl, unsigned long *xr);
+void Blowfish_Encrypt(BLOWFISH_CTX *ctx, unsigned int *xl, unsigned int *xr);
+void Blowfish_Decrypt(BLOWFISH_CTX *ctx, unsigned int *xl, unsigned int *xr);
 
 #ifdef __cplusplus
 };
